tecdoc-caching-app/requirements.txt
Flask
python-dotenv
apify-client
psycopg2-binary  # Si usas PostgreSQL
